+++ packages/vats/src/core/boot-sim.js
@@ -2,29 +2,23 @@
 import { makeBootstrap } from './lib-boot.js';
 
 import * as basicBehaviorsPlus from './basic-behaviors.js';
-import { SHARED_CHAIN_BOOTSTRAP_MANIFEST } from './chain-behaviors.js';
-import { SIM_CHAIN_BOOTSTRAP_PERMITS } from './sim-behaviors.js';
 import * as chainBehaviorsPlus from './chain-behaviors.js';
 import * as simBehaviorsPlus from './sim-behaviors.js';
 import * as utils from './utils.js';
+
+const { BASIC_BOOTSTRAP_PERMITS: _b, ...basicBehaviors } = basicBehaviorsPlus;
+const {
+  CHAIN_BOOTSTRAP_MANIFEST: _c,
+  SHARED_CHAIN_BOOTSTRAP_MANIFEST,
+  ...chainBehaviors
+} = chainBehaviorsPlus;
+const { SIM_CHAIN_BOOTSTRAP_PERMITS, ...simBehaviors } = simBehaviorsPlus;
 
 export const MANIFEST = {
   ...SHARED_CHAIN_BOOTSTRAP_MANIFEST,
   ...SIM_CHAIN_BOOTSTRAP_PERMITS,
 };
 
-const {
-  BASIC_BOOTSTRAP_PERMITS: _b,
-  PowerFlags: _p,
-  makeMyAddressNameAdminKit: _m,
-  ...basicBehaviors
-} = basicBehaviorsPlus;
-const {
-  CHAIN_BOOTSTRAP_MANIFEST: _c,
-  SHARED_CHAIN_BOOTSTRAP_MANIFEST: _s,
-  ...chainBehaviors
-} = chainBehaviorsPlus;
-const { SIM_CHAIN_BOOTSTRAP_PERMITS: _sc, ...simBehaviors } = simBehaviorsPlus;
 const behaviors = { ...basicBehaviors, ...chainBehaviors, ...simBehaviors };
 
 const modules = harden({ behaviors: { ...behaviors }, utils: { ...utils } });
